Move settings facade to shared
