--- src/farm/modules/settings/services/index.js
@@ -1,2 +1,0 @@
-export { SettingsService } from './SettingsService';
-export { LocalSettingsService } from './LocalSettingsService';
