+++ src/shared/settings/services/LocalSettingsService.js
