+++ src/shared/settings/services/index.js
@@ -1,0 +1,2 @@
+export { GlobalSettingsService } from './GlobalSettingsService';
+export { LocalSettingsService } from './LocalSettingsService';
